Refactor loader
