+++ src/_utils/templateWithLoader.js
@@ -1,0 +1,15 @@
+import templateWithHoc from './templateWithHoc'
+import { clearCommentsRgx } from './constants'
+import hasExportName from './hasExportName'
+import hasHOC from './hasHOC'
+
+function templateWithLoader(
+  rawCode,
+  { i18nFile, arePagesInsideSrc, page, prefix, loader } = {}
+) {
+  console.log('@todo templateWithLoader', loader)
+
+  return rawCode // @todo implement
+}
+
+export default templateWithLoader
